Resolve secp256k1_ecdsa_verify issue with TX validation.
Non-segwit transactions (currently all transactions) are now hashed using the legacy code from the previous Cloak codebase. This ensures a matching TX comparison hash is always generated for use with TX signature hash validation.

It is doubtful that this code will need to change in future, but if it is decided that the verificaiion method needs to change, then a network fork will need  to be performed so that newer blocks use a new TX signature hash.